resources/requirements.txt
PyYAML
jsonschema
jinja2
jsonschema